add data validation to new folders and documents

+++ backend/app/routers/workspace_manager.py
@@ -6,7 +6,8 @@ import json
 import os
 from datetime import datetime
 from pathlib import Path
-from fastapi import APIRouter, Depends, HTTPException
+from re import fullmatch
+from fastapi import APIRouter, Depends, HTTPException, status
 
 from app.routers.auth import get_current_user
 from app.utils.message import Message, MsgStatus
@@ -48,7 +49,7 @@ def get_workspace_path(workspace_name: str = "") -> Path:
     path = Path(".") / DATA_DIR / WORKSPACES_DIR / workspace_name
     if not path.exists():
         raise HTTPException(
-            status_code=404,
+            status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Workspace with name <<{workspace_name}>> doesn't exist",
         )
 
@@ -82,9 +83,9 @@ def get_document_path(workspace_name: str, doc_name: str = "") -> Path:
     path = workspace_path / DOCUMENTS_DIR / doc_name
     if not path.exists():
         raise HTTPException(
-            status_code=404,
+            status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Document with name <<{doc_name}>>"
-            + f" doesn't exist in workspace <<{workspace_name}>>",
+                   + f" doesn't exist in workspace <<{workspace_name}>>",
         )
 
     return path
@@ -118,7 +119,7 @@ def get_image_path(workspace_name: str, document_name: str, img: str = "") -> Pa
     path = document_path / IMAGES_DIR / img
     if not path.exists():
         raise HTTPException(
-            status_code=404,
+            status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Image with name <<{img}>> doesn't exist in workspace <<{workspace_name}>>",
         )
 
@@ -126,7 +127,7 @@ def get_image_path(workspace_name: str, document_name: str, img: str = "") -> Pa
 
 
 def get_attachment_path(
-    workspace_name: str, document_name: str, atch: str = ""
+        workspace_name: str, document_name: str, atch: str = ""
 ) -> Path:
     """
     Returns path to attachment file with given name.
@@ -155,23 +156,24 @@ def get_attachment_path(
     path = document_path / ATTACHMENTS_DIR / atch
     if not path.exists():
         raise HTTPException(
-            status_code=404,
+            status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Attachment with name <<{atch}>>"
-            + f" doesn't exist in workspace <<{workspace_name}>>",
+                   + f" doesn't exist in workspace <<{workspace_name}>>",
         )
 
     return path
 
 
 async def add_document_to_virtual_structure(
-    workspace_name: str, document_name: str, virtual_path: str
+        workspace_name: str, document_name: str, virtual_path: str
 ) -> Message:
     """TODO function docstring"""
 
     path = get_workspace_path(workspace_name) / CONFIG_FILE
     if not path.exists():
         raise HTTPException(
-            status_code=404, detail=f"Can't find config file at {path.absolute()}"
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Can't find config file at {path.absolute()}"
         )
 
     with open(path, "r") as config_file:
@@ -185,6 +187,87 @@ async def add_document_to_virtual_structure(
             break
     else:
         config_data["virtual_structure"].append(value)
+
+    with open(path, "w") as config_file:
+        json.dump(config_data, config_file, indent=4)
+
+    return Message(
+        status=MsgStatus.INFO,
+        detail=f"<<{workspace_name}>> virtual structure updated successfully",
+    )
+
+
+def clear_directory(path: Path):
+    """TODO function docstring"""
+
+    for child in path.glob("*"):
+        if child.is_file():
+            child.unlink()
+        else:
+            clear_directory(child)
+    path.rmdir()
+
+
+@router.post("/{workspace_name}/new_folder/{folder_name}")
+async def add_folder_to_virtual_structure(
+        workspace_name: str, folder_name: str, virtual_path: str
+) -> Message:
+    """TODO function docstring"""
+
+    if not fullmatch("[A-Za-z0-9-_]+", folder_name):
+        raise HTTPException(
+            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
+            detail="Folder name must contain only upper and lower case characters, "
+                   + "numbers, and symbols - or _ ",
+        )
+
+    path = get_workspace_path(workspace_name) / CONFIG_FILE
+    if not path.exists():
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Can't find config file at {path.absolute()}"
+        )
+
+    with open(path, "r") as config_file:
+        config_data = json.load(config_file)
+
+    value = {"name": folder_name, "type": "folder", "virtual_path": virtual_path}
+
+    for element in config_data["virtual_structure"]:
+        if element["name"] == folder_name and element["type"] == "folder":
+            element["virtual_path"] = virtual_path
+            break
+    else:
+        config_data["virtual_structure"].append(value)
+
+    with open(path, "w") as config_file:
+        json.dump(config_data, config_file, indent=4)
+
+    return Message(
+        status=MsgStatus.INFO,
+        detail=f"<<{workspace_name}>> virtual structure updated successfully",
+    )
+
+
+@router.post("/{workspace_name}/remove_folder/{folder_name}")
+async def remove_folder_from_virtual_structure(
+        workspace_name: str, folder_name: str
+) -> Message:
+    """TODO function docstring"""
+
+    path = get_workspace_path(workspace_name) / CONFIG_FILE
+    if not path.exists():
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Can't find config file at {path.absolute()}"
+        )
+
+    with open(path, "r") as config_file:
+        config_data = json.load(config_file)
+
+    for element in config_data["virtual_structure"]:
+        if element["name"] == folder_name and element["type"] == "folder":
+            config_data["virtual_structure"].remove(element)
 
     with open(path, "w") as config_file:
         json.dump(config_data, config_file, indent=4)
@@ -223,78 +306,6 @@ def _add_new_folder(folder: dict, new_folder_name: str, level: int):
     return new_folder, level + 1
 
 
-def clear_directory(path: Path):
-    """TODO function docstring"""
-
-    for child in path.glob("*"):
-        if child.is_file():
-            child.unlink()
-        else:
-            clear_directory(child)
-    path.rmdir()
-
-
-@router.post("/{workspace_name}/new_folder/{folder_name}")
-async def add_folder_to_virtual_structure(
-    workspace_name: str, folder_name: str, virtual_path: str
-) -> Message:
-    """TODO function docstring"""
-
-    path = get_workspace_path(workspace_name) / CONFIG_FILE
-    if not path.exists():
-        raise HTTPException(
-            status_code=404, detail=f"Can't find config file at {path.absolute()}"
-        )
-
-    with open(path, "r") as config_file:
-        config_data = json.load(config_file)
-
-    value = {"name": folder_name, "type": "folder", "virtual_path": virtual_path}
-
-    for element in config_data["virtual_structure"]:
-        if element["name"] == folder_name and element["type"] == "folder":
-            element["virtual_path"] = virtual_path
-            break
-    else:
-        config_data["virtual_structure"].append(value)
-
-    with open(path, "w") as config_file:
-        json.dump(config_data, config_file, indent=4)
-
-    return Message(
-        status=MsgStatus.INFO,
-        detail=f"<<{workspace_name}>> virtual structure updated successfully",
-    )
-
-
-@router.post("/{workspace_name}/remove_folder/{folder_name}")
-async def remove_folder_from_virtual_structure(
-    workspace_name: str, folder_name: str
-) -> Message:
-    """TODO function docstring"""
-
-    path = get_workspace_path(workspace_name) / CONFIG_FILE
-    if not path.exists():
-        raise HTTPException(
-            status_code=404, detail=f"Can't find config file at {path.absolute()}"
-        )
-
-    with open(path, "r") as config_file:
-        config_data = json.load(config_file)
-
-    for element in config_data["virtual_structure"]:
-        if element["name"] == folder_name and element["type"] == "folder":
-            config_data["virtual_structure"].remove(element)
-
-    with open(path, "w") as config_file:
-        json.dump(config_data, config_file, indent=4)
-
-    return Message(
-        status=MsgStatus.INFO,
-        detail=f"<<{workspace_name}>> virtual structure updated successfully",
-    )
-
-
 @router.get("/structure/tree/{workspace_name}")
 async def get_workspace_tree_structure(workspace_name: str) -> json:
     """
@@ -314,7 +325,8 @@ async def get_workspace_tree_structure(workspace_name: str) -> json:
     path = get_workspace_path(workspace_name) / CONFIG_FILE
     if not path.exists():
         raise HTTPException(
-            status_code=404, detail=f"Can't find config file at {path.absolute()}"
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Can't find config file at {path.absolute()}"
         )
 
     with open(path, "r") as config_file:
@@ -370,7 +382,8 @@ async def get_workspace_raw_structure(workspace_name: str) -> json:
     path = get_workspace_path(workspace_name) / CONFIG_FILE
     if not path.exists():
         raise HTTPException(
-            status_code=404, detail=f"Can't find config file at {path.absolute()}"
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Can't find config file at {path.absolute()}"
         )
 
     with open(path, "r") as config_file:
@@ -381,7 +394,7 @@ async def get_workspace_raw_structure(workspace_name: str) -> json:
 
 @router.post("/new/{workspace_name}", response_model=Message, status_code=201)
 async def create_new_workspace(
-    workspace_name: str, private: bool, creator: str = Depends(get_current_user)
+        workspace_name: str, private: bool, creator: str = Depends(get_current_user)
 ) -> Message:
     """
     Create a new workspace and its subdirectories
@@ -397,13 +410,21 @@ async def create_new_workspace(
     Raises:
         HTTPException (409): if selected name collides with already
                              existing workspace
+        HTTPException (422): if selected name is incorrect
     """
+
+    if not fullmatch("[A-Za-z0-9-_]+", workspace_name):
+        raise HTTPException(
+            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
+            detail="Workspace name must contain only upper and lower case characters, "
+                   + "numbers, and symbols - or _ ",
+        )
 
     path = get_workspace_path() / workspace_name
 
     if path.exists():
         raise HTTPException(
-            status_code=409,
+            status_code=status.HTTP_409_CONFLICT,
             detail=f"Workspace with name <<{workspace_name}>> already exists",
         )
 
@@ -435,7 +456,7 @@ async def create_new_workspace(
 
 @router.post("/remove/{workspace_name}", response_model=Message, status_code=200)
 async def remove_workspace(
-    workspace_name: str, user: str = Depends(get_current_user)
+        workspace_name: str, user: str = Depends(get_current_user)
 ) -> Message:
     """TODO function docstring"""
 
@@ -446,7 +467,7 @@ async def remove_workspace(
 
     if user["username"] != info_data["creator"]:
         raise HTTPException(
-            status_code=401,
+            status_code=status.HTTP_401_UNAUTHORIZED,
             detail=f"Only creator of the workspace - {info_data['creator']}, can delete it",
         )
 
@@ -459,10 +480,10 @@ async def remove_workspace(
     "/new/{workspace_name}/{document_name}", response_model=Message, status_code=201
 )
 async def create_new_document(
-    workspace_name: str,
-    document_name: str,
-    virtual_path: str,
-    creator: str = Depends(get_current_user),
+        workspace_name: str,
+        document_name: str,
+        virtual_path: str,
+        creator: str = Depends(get_current_user),
 ) -> Message:
     """
     Create a new document in given workspace
@@ -470,8 +491,8 @@ async def create_new_document(
     Parameters:
         workspace_name (str): name of the workspace that contains the document
         document_name (str): name of the new document
-        virtual_path (str): todo
-        creator (str): todo
+        virtual_path (str): path of the new document
+        creator (str): username of the creator of the new document
 
     Returns:
         Message: details about result of operation
@@ -479,13 +500,21 @@ async def create_new_document(
     Raises:
         HTTPException (409): if selected name collides with already
                              existing document
+        HTTPException (422): if selected name is incorrect
     """
+
+    if not fullmatch("[A-Za-z0-9-_]+", document_name):
+        raise HTTPException(
+            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
+            detail="Document name must contain only upper and lower case characters, "
+                   + "numbers, and symbols - or _ ",
+        )
 
     path = get_workspace_path(workspace_name) / DOCUMENTS_DIR / document_name
 
     if path.exists():
         raise HTTPException(
-            status_code=409,
+            status_code=status.HTTP_409_CONFLICT,
             detail=f"Document <<{document_name}>> already exists in workspace <<{workspace_name}>>",
         )
 
@@ -510,7 +539,7 @@ async def create_new_document(
     "/remove/{workspace_name}/{document_name}", response_model=Message, status_code=200
 )
 async def remove_document(
-    workspace_name: str, document_name: str, user: str = Depends(get_current_user)
+        workspace_name: str, document_name: str, user: str = Depends(get_current_user)
 ) -> Message:
     """TODO function docstring"""
 
@@ -521,7 +550,8 @@ async def remove_document(
 
     if user["username"] != info_data["creator"]:
         raise HTTPException(
-            status_code=401, detail="Only creator of the workspace can delete it"
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Only creator of the workspace can delete it"
         )
 
     path = get_workspace_path(workspace_name) / CONFIG_FILE
@@ -555,7 +585,7 @@ async def load_document_content(workspace_name: str, document_name: str) -> json
 
 @router.post("/{workspace_name}/{document_name}")
 async def save_document_content(
-    workspace_name: str, document_name: str, document_data: list
+        workspace_name: str, document_name: str, document_data: list
 ) -> Message:
     """TODO function docstring"""
